Implement auto-upgrade on demand
