+++ engine/launcher/src/main/scala/org/enso/launcher/components/Manifest.scala
@@ -6,7 +6,7 @@ import java.nio.file.Path
 import cats.Show
 import io.circe.{yaml, Decoder}
 import nl.gn0s1s.bump.SemVer
-import org.enso.launcher.OS
+import org.enso.launcher.{CurrentVersion, OS}
 import org.enso.launcher.components.Manifest.JVMOption
 import org.enso.pkg.SemVerJson._
 
@@ -41,6 +41,9 @@ case class Manifest(
     */
   def runtimeVersion: RuntimeVersion =
     RuntimeVersion(graalVMVersion, graalJavaVersion)
+
+  def isUsableWithCurrentVersion: Boolean =
+    CurrentVersion.version >= minimumLauncherVersion
 }
 
 object Manifest {
