+++ project/NativeImage.scala
@@ -168,7 +168,8 @@ object NativeImage {
     else file(name)
 
   private val muslBundleUrl =
-    "https://github.com/gradinac/musl-bundle-example/releases/download/v1.0/musl.tar.gz"
+    "https://github.com/gradinac/musl-bundle-example/releases/download/" +
+    "v1.0/musl.tar.gz"
 
   /**
     * Ensures that the `musl` bundle is installed.
